Refactor configuration management and refactored HexStrikeCache
