+++ config.py
@@ -5,7 +5,9 @@ from typing import Any, Optional
 _config = {
     "APP_NAME": "HexStrike AI Community Edition",
     "VERSION": "1.0.4",
-    "COMMAND_TIMEOUT": 300
+    "COMMAND_TIMEOUT": 300,
+    "CACHE_SIZE": 1000,
+    "CACHE_TTL": 3600  # 1 hour
 }
 
 def get(key: str, default: Optional[Any] = None) -> Any:
